+++ processor/src/test/resources/expected/com/example/component/Arez_PublicCtorNonPublicModel.java
@@ -1,0 +1,157 @@
+package com.example.component;
+
+import arez.Arez;
+import arez.ArezContext;
+import arez.Component;
+import arez.Disposable;
+import arez.component.ComponentState;
+import arez.component.DisposeNotifier;
+import arez.component.DisposeTrackable;
+import arez.component.Identifiable;
+import javax.annotation.Generated;
+import javax.annotation.Nonnull;
+import javax.annotation.Nullable;
+import org.realityforge.braincheck.Guards;
+
+@Generated("arez.processor.ArezProcessor")
+final class Arez_PublicCtorNonPublicModel extends PublicCtorNonPublicModel implements Disposable, Identifiable<Integer>, DisposeTrackable {
+  private static volatile int $$arezi$$_nextId;
+
+  private final int $$arezi$$_id;
+
+  private byte $$arezi$$_state;
+
+  @Nullable
+  private final ArezContext $$arezi$$_context;
+
+  private final Component $$arezi$$_component;
+
+  private final DisposeNotifier $$arezi$$_disposeNotifier;
+
+  Arez_PublicCtorNonPublicModel() {
+    super();
+    this.$$arezi$$_context = Arez.areZonesEnabled() ? Arez.context() : null;
+    this.$$arezi$$_id = ( Arez.areNamesEnabled() || Arez.areRegistriesEnabled() || Arez.areNativeComponentsEnabled() ) ? $$arezi$$_nextId++ : 0;
+    if ( Arez.shouldCheckApiInvariants() ) {
+      this.$$arezi$$_state = ComponentState.COMPONENT_INITIALIZED;
+    }
+    this.$$arezi$$_component = Arez.areNativeComponentsEnabled() ? $$arezi$$_context().component( "PublicCtorNonPublicModel", $$arezi$$_id(), Arez.areNamesEnabled() ? $$arezi$$_name() : null, () -> $$arezi$$_preDispose() ) : null;
+    this.$$arezi$$_disposeNotifier = new DisposeNotifier();
+    if ( Arez.shouldCheckApiInvariants() ) {
+      this.$$arezi$$_state = ComponentState.COMPONENT_CONSTRUCTED;
+    }
+    if ( Arez.areNativeComponentsEnabled() ) {
+      this.$$arezi$$_component.complete();
+    }
+    if ( Arez.shouldCheckApiInvariants() ) {
+      this.$$arezi$$_state = ComponentState.COMPONENT_READY;
+    }
+  }
+
+  final ArezContext $$arezi$$_context() {
+    if ( Arez.shouldCheckApiInvariants() ) {
+      Guards.apiInvariant( () -> ComponentState.hasBeenInitialized( this.$$arezi$$_state ), () -> "Method named '$$arezi$$_context' invoked on uninitialized component of type 'PublicCtorNonPublicModel'" );
+    }
+    return Arez.areZonesEnabled() ? this.$$arezi$$_context : Arez.context();
+  }
+
+  final int $$arezi$$_id() {
+    if ( Arez.shouldCheckInvariants() && !Arez.areNamesEnabled() && !Arez.areRegistriesEnabled() && !Arez.areNativeComponentsEnabled() ) {
+      Guards.fail( () -> "Method invoked to access id when id not expected on component named '" + $$arezi$$_name() + "'." );
+    }
+    return this.$$arezi$$_id;
+  }
+
+  @Override
+  @Nonnull
+  public final Integer getArezId() {
+    return $$arezi$$_id();
+  }
+
+  String $$arezi$$_name() {
+    if ( Arez.shouldCheckApiInvariants() ) {
+      Guards.apiInvariant( () -> ComponentState.hasBeenInitialized( this.$$arezi$$_state ), () -> "Method named '$$arezi$$_name' invoked on uninitialized component of type 'PublicCtorNonPublicModel'" );
+    }
+    return "PublicCtorNonPublicModel." + $$arezi$$_id();
+  }
+
+  private void $$arezi$$_preDispose() {
+    $$arezi$$_disposeNotifier.dispose();
+  }
+
+  @Override
+  @Nonnull
+  public DisposeNotifier getNotifier() {
+    return $$arezi$$_disposeNotifier;
+  }
+
+  @Override
+  public boolean isDisposed() {
+    return ComponentState.isDisposingOrDisposed( this.$$arezi$$_state );
+  }
+
+  @Override
+  public void dispose() {
+    if ( !ComponentState.isDisposingOrDisposed( this.$$arezi$$_state ) ) {
+      this.$$arezi$$_state = ComponentState.COMPONENT_DISPOSING;
+      if ( Arez.areNativeComponentsEnabled() ) {
+        this.$$arezi$$_component.dispose();
+      } else {
+        $$arezi$$_context().safeAction( Arez.areNamesEnabled() ? $$arezi$$_name() + ".dispose" : null, true, false, () -> { {
+          this.$$arezi$$_preDispose();
+        } } );
+      }
+      if ( Arez.shouldCheckApiInvariants() ) {
+        this.$$arezi$$_state = ComponentState.COMPONENT_DISPOSED;
+      }
+    }
+  }
+
+  @Override
+  public void doStuff(final long time, final float someOtherParameter) {
+    if ( Arez.shouldCheckApiInvariants() ) {
+      Guards.apiInvariant( () -> ComponentState.isActive( this.$$arezi$$_state ), () -> "Method named 'doStuff' invoked on " + ComponentState.describe( this.$$arezi$$_state ) + " component named '" + $$arezi$$_name() + "'" );
+    }
+    try {
+      $$arezi$$_context().safeAction(Arez.areNamesEnabled() ? $$arezi$$_name() + ".doStuff" : null, true, false, false, () -> super.doStuff(time,someOtherParameter), time, someOtherParameter );
+    } catch( final RuntimeException | Error $$arez_exception$$ ) {
+      throw $$arez_exception$$;
+    } catch( final Throwable $$arez_exception$$ ) {
+      throw new IllegalStateException( $$arez_exception$$ );
+    }
+  }
+
+  @Override
+  public final int hashCode() {
+    if ( Arez.areNativeComponentsEnabled() ) {
+      return Integer.hashCode( $$arezi$$_id() );
+    } else {
+      return super.hashCode();
+    }
+  }
+
+  @Override
+  public final boolean equals(final Object o) {
+    if ( Arez.areNativeComponentsEnabled() ) {
+      if ( this == o ) {
+        return true;
+      } else if ( null == o || !(o instanceof Arez_PublicCtorNonPublicModel) ) {
+        return false;
+      } else {
+        final Arez_PublicCtorNonPublicModel that = (Arez_PublicCtorNonPublicModel) o;;
+        return $$arezi$$_id() == that.$$arezi$$_id();
+      }
+    } else {
+      return super.equals( o );
+    }
+  }
+
+  @Override
+  public final String toString() {
+    if ( Arez.areNamesEnabled() ) {
+      return "ArezComponent[" + $$arezi$$_name() + "]";
+    } else {
+      return super.toString();
+    }
+  }
+}