+++ processor/src/test/resources/input/com/example/component/PublicCtorNonPublicModel.java
@@ -1,0 +1,17 @@
+package com.example.component;
+
+import arez.annotations.Action;
+import arez.annotations.ArezComponent;
+
+@ArezComponent
+abstract class PublicCtorNonPublicModel
+{
+  public PublicCtorNonPublicModel()
+  {
+  }
+
+  @Action
+  public void doStuff( final long time, float someOtherParameter )
+  {
+  }
+}